+++ forge/src/main/java/xyz/faewulf/backpack/RyoamicDynLights/RyoamicDynLightsModule.java
@@ -1,0 +1,7 @@
+package xyz.faewulf.backpack.RyoamicDynLights;
+
+import xyz.faewulf.backpack.inter.IDynamicLightCompatLayer;
+
+public class RyoamicDynLightsModule implements IDynamicLightCompatLayer {
+
+}