+++ forge/src/main/java/xyz/faewulf/backpack/event/onClientInit.java
@@ -1,0 +1,14 @@
+package xyz.faewulf.backpack.event;
+
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import xyz.faewulf.backpack.Constants;
+
+@Mod.EventBusSubscriber(modid = Constants.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+public class onClientInit {
+    @SubscribeEvent
+    public static void onClientInitiation(FMLClientSetupEvent event) {
+    }
+}