+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-# Changelog v1.0.0
+# Changelog v1.1.0
 
 ### Added
 
+- Ported to forge mod loader.
+
 ### Changed
 
-### Removed
+- Only works with `traverlers backpack` version `10.1.2x`.
 
 ### Fixed
+
+- Compatible with `traverlers backpack` version `10.1.2x`